requirements.txt
albumentations==1.4.3
braceexpand==0.1.7
bytecode==0.15.1
diffusers==0.27.2
easydict==1.13
Flask==3.0.3
gradio==4.26.0
imageio==2.34.0
imageio_ffmpeg==0.4.9
imgaug==0.4.0
joblib==1.4.0
jpeg4py==0.1.4
kornia==0.7.2
lmdb==1.4.1
MarkupSafe==2.1.5
matplotlib==3.8.4
memory_profiler==0.61.0
numpy==1.26.4
omegaconf==2.3.0
onnxruntime==1.17.1
pandas==2.2.2
Pillow==10.3.0
pycocotools==2.0.7
pytorch_lightning==2.2.1
PyYAML==6.0.1
scikit_learn==1.4.2
scipy==1.13.0
setuptools==65.5.0
setuptools==69.2.0
six==1.16.0
skimage==0.0
sync_batchnorm==0.0.1
tensorboardX==2.6.2.2
tensorflow==2.16.1
thop==0.1.1.post2209072238
timm==0.9.16
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
webdataset==0.2.86
